requirements.txt
python-dotenv
requests
openai
numpy
pandas
